add test (failing)

+++ src/web/peers.cpp
@@ -1,6 +1,7 @@
 #include "peers.hpp"
 
 #include <bits/stdint-intn.h>
+#include <bits/stdint-uintn.h>
 #include <ctime>
 #include <uvw.hpp>
 
@@ -25,7 +26,7 @@ void printHex(const char* toPrint, int size, int width = 16) {
         if (i % width == 0) {
             std::cout << std::endl;
         }
-        std::cout << std::hex << std::setfill('0') << std::setw(2) << (int) toPrint[i] << ' ';
+        std::cout << std::hex << std::setfill('0') << std::setw(2) << (uint8_t) toPrint[i] << ' ';
     }
     std::cout << std::endl << std::dec;
 }
